feat: 송신 쪽지 리스트 조회 추가

See merge request s11-final/S11P31B104!10

+++ star/src/main/java/com/ssafy/star/message/entity/Message.java
@@ -29,7 +29,7 @@ public class Message {
     private Member sender;
 
     @Column(name="receiver_type", nullable = false)
-    private int receiverType = 0;
+    private short receiverType = 0;
 
     @Column(name = "hint_image_first")
     private String hintImageFirst;
